+++ reparse/reparse.mli
@@ -1,0 +1,2 @@
+include module type of Parser
+include module type of Parser.Let_syntax