+++ CHANGES.md
@@ -1,6 +1,6 @@
-## v3.0.0 2021-01-14
+## v3.0.0 2021-01- [unreleased]
 
-This release has backwards incompatible changes. 
+This release has backwards incompatible changes.
 
 - Infix functions are now available in `Parser` moduel itself.
 - Add support for let operators `and+` and `and*`.
@@ -14,6 +14,7 @@ This release has backwards incompatible changes.
   conventions in `base` library dependency.
 - Removed `map4` function.
 - Add support for `ppx_let`.
+- Deprecate `Parser` module. Use `Reparse` instead.
 
 ## v2.0.0 2020-11-09 UK